extracted construct_service.go
